RPM-Meter/requirements.txt
asyncio
websockets
RPi.GPIO